Refactor semi-implicit timesteppers
The previous implementation assumed the linear part was specified by a matrix,
and the solver needed for semi-implicit timesteppers was determined by LU
decomposition of this matrix.

This can be inefficient in some common cases, for instance if the linear part is
diagonal.  Then the solve is easily done without the need for LU decomposition.

In the new implementation, the linear part is given more generally by a
function, and the user can specify a solver to use for the semi-implicit
methods.  One may still specify the linear part by matrix (via the attribute
_linear), in which case the user does not need to specify a solver for the
semi-implicit timestepper, and the LU decomposition is used, as before.