Refactor Model class and add type hints
Previously, the Model class contained a method `set_stepper` which re-defined
other methods (`step` and `adjoint_step`) within the Model object.  This was
confusing and made the Model class too complex.

In this version, the method `set_stepper` is replaced with a method
`get_stepper` which returns a callable object that advances the state forward by
one timestep.  There is also an analogous method `get_adjoint_stepper`.

The examples and unit tests are updated to use the new interface.

In addition, some type hints have been added.