Add MultiLinear, abstract AE, and AEList class
The MultiLinear class defines a projection-based multilinear
network. Unlike the IRMAE paper, the decoding layers here are
inverses of the encoding layers. The abstract AE class is an
abstract class defining the structure of a romnet autoencoder.
The AEList class defines an AE generated by smaller subcomponent
autoencoders. This is useful when defining an autoencoder
structure like Convolutional + AE + MultiLinear or
AE + MultiLinear.